Metadata of Medical Images
Medical Image Metadata Description
The bone image files were read into MATLAB for computational analysis. Each image contains the following metadata:

Source: MEDPIX
Image Width: The width of the image in pixels
Image Height: The height of the image in pixels
Image Name: The name of the image file, including the .jpg extension
Regions of Interest (ROIs): The number of regions of interest in the image
Medical Image Type: The type of medical image (e.g., X-ray, CT, MRI)
Sub-regions of Interest (SubROIs): The number of sub-regions of interest within each region
This information is used in the analysis of the bone images to extract relevant features and insights.